requirements.txt
requests
psycopg2-binary
colorama
Flask  # Thêm dòng này
